demos/adk-with-skills/banana_ric_agent/requirements.txt
google-genai
Pillow
google-adk
google-cloud-aiplatform
google-cloud-storage
